Added Timer plus..
Added global timer, timer text and code to globalFunctions,
added code to launch DeathScene with the others and sleep it.  This so we can kill the others at the right time without the anoying flashes of their maps.
added death by volcano message to deathScene if volcano blew.
Things left:
sound and volcano shake and blow if we have the time.